Remove SCSS and fix console warnings
I removed the commented SCSS as there was no point in keeping it.
I missed the authentication.styles which have now been converted to a styled-component.

I converted props being used in styled-components to 'transient props' by prefixing the prop names with '$'.
This stops it from being passed to React node or rendered on the DOM i.e. just used in the styled-component.